--- a/backend/open_webui/models/prompts.py
+++ b/backend/open_webui/models/prompts.py
@@ -118,18 +118,191 @@ class PromptsTable:
 
             return prompts
 
-    def get_prompts_by_user_id(
-        self, user_id: str, permission: str = "write"
-    ) -> list[PromptUserResponse]:
-        prompts = self.get_prompts()
+    def get_prompts_paginated(
+        self, page: int = 1, limit: int = 20, search: str = None
+    ) -> list[PromptModel]:
+        """Get paginated prompts with optional search"""
+        with get_db() as db:
+            query = db.query(Prompt)
 
-        return [
-            prompt
-            for prompt in prompts
-            if prompt.user_id == user_id
-            or prompt.access_control == None
-            or has_access(user_id, permission, prompt.access_control)
-        ]
+            # Apply search filter if provided
+            if search and search.strip():
+                search_term = f"%{search.strip().lower()}%"
+                query = query.filter(
+                    Prompt.command.ilike(search_term)
+                    | Prompt.title.ilike(search_term)
+                    | Prompt.content.ilike(search_term)
+                )
+
+            # Apply pagination
+            offset = (page - 1) * limit
+            prompts = (
+                query.order_by(Prompt.timestamp.desc())
+                .offset(offset)
+                .limit(limit)
+                .all()
+            )
+
+            return [PromptModel.model_validate(prompt) for prompt in prompts]
+
+    def get_prompts_with_users_paginated(
+        self, page: int = 1, limit: int = 20, search: str = None
+    ) -> list[PromptUserResponse]:
+        """Get paginated prompts with user info and optional search"""
+        with get_db() as db:
+            query = db.query(Prompt)
+
+            # Apply search filter if provided
+            if search and search.strip():
+                search_term = f"%{search.strip().lower()}%"
+                query = query.filter(
+                    Prompt.command.ilike(search_term)
+                    | Prompt.title.ilike(search_term)
+                    | Prompt.content.ilike(search_term)
+                )
+
+            # Apply pagination
+            offset = (page - 1) * limit
+            prompts_data = (
+                query.order_by(Prompt.timestamp.desc())
+                .offset(offset)
+                .limit(limit)
+                .all()
+            )
+
+            prompts = []
+            for prompt in prompts_data:
+                user = Users.get_user_by_id(prompt.user_id)
+                prompts.append(
+                    PromptUserResponse.model_validate(
+                        {
+                            **PromptModel.model_validate(prompt).model_dump(),
+                            "user": user.model_dump() if user else None,
+                        }
+                    )
+                )
+
+            return prompts
+
+    def get_prompts_count(self, search: str = None) -> int:
+        """Get total count of prompts with optional search filter"""
+        with get_db() as db:
+            query = db.query(Prompt)
+
+            # Apply search filter if provided
+            if search and search.strip():
+                search_term = f"%{search.strip().lower()}%"
+                query = query.filter(
+                    Prompt.command.ilike(search_term)
+                    | Prompt.title.ilike(search_term)
+                    | Prompt.content.ilike(search_term)
+                )
+
+            return query.count()
+
+    def get_prompts_by_user_id_paginated(
+        self,
+        user_id: str,
+        permission: str = "write",
+        page: int = 1,
+        limit: int = 20,
+        search: str = None,
+    ) -> list[PromptModel]:
+        """Get paginated prompts by user with permission check and optional search"""
+        # For pagination, we need to filter in the database for efficiency
+        with get_db() as db:
+            query = db.query(Prompt)
+
+            # Apply search filter if provided
+            if search and search.strip():
+                search_term = f"%{search.strip().lower()}%"
+                query = query.filter(
+                    Prompt.command.ilike(search_term)
+                    | Prompt.title.ilike(search_term)
+                    | Prompt.content.ilike(search_term)
+                )
+
+            # Filter by user access: public prompts, user's own prompts, or prompts with specific access
+            query = query.filter(
+                (Prompt.user_id == user_id)  # User's own prompts
+                | (Prompt.access_control.is_(None))  # Public prompts
+                # Note: For complex access control with groups, we'd need additional logic
+            )
+
+            # Apply pagination
+            offset = (page - 1) * limit
+            prompts = (
+                query.order_by(Prompt.timestamp.desc())
+                .offset(offset)
+                .limit(limit)
+                .all()
+            )
+
+            # Filter by access control for the remaining prompts
+            result = []
+            for prompt_data in prompts:
+                prompt = PromptModel.model_validate(prompt_data)
+                if (
+                    prompt.user_id == user_id
+                    or prompt.access_control is None
+                    or has_access(user_id, permission, prompt.access_control)
+                ):
+                    result.append(prompt)
+
+            return result
+
+    def get_prompts_by_user_id_with_users_paginated(
+        self,
+        user_id: str,
+        permission: str = "write",
+        page: int = 1,
+        limit: int = 20,
+        search: str = None,
+    ) -> list[PromptUserResponse]:
+        """Get paginated prompts by user with user info, permission check and optional search"""
+        # Get prompts first
+        prompts = self.get_prompts_by_user_id_paginated(
+            user_id, permission, page, limit, search
+        )
+
+        # Add user info
+        result = []
+        for prompt in prompts:
+            user = Users.get_user_by_id(prompt.user_id)
+            result.append(
+                PromptUserResponse.model_validate(
+                    {
+                        **prompt.model_dump(),
+                        "user": user.model_dump() if user else None,
+                    }
+                )
+            )
+
+        return result
+
+    def get_prompts_count_by_user_id(
+        self, user_id: str, permission: str = "write", search: str = None
+    ) -> int:
+        """Get total count of prompts accessible by user with optional search filter"""
+        with get_db() as db:
+            query = db.query(Prompt)
+
+            # Apply search filter if provided
+            if search and search.strip():
+                search_term = f"%{search.strip().lower()}%"
+                query = query.filter(
+                    Prompt.command.ilike(search_term)
+                    | Prompt.title.ilike(search_term)
+                    | Prompt.content.ilike(search_term)
+                )
+
+            # Filter by user access: public prompts, user's own prompts, or prompts with specific access
+            query = query.filter(
+                (Prompt.user_id == user_id)  # User's own prompts
+                | (Prompt.access_control.is_(None))  # Public prompts
+            )
+
+            return query.count()
 
     def update_prompt_by_command(
         self, command: str, form_data: PromptForm