--- a/src/lib/apis/evaluations/index.ts
+++ b/src/lib/apis/evaluations/index.ts
@@ -62,7 +62,88 @@ export const updateConfig = async (token: string, config: object) => {
 	return res;
 };
 
-export const getAllFeedbacks = async (token: string = '') => {
+export const getAllFeedbacks = async (
+	token: string = '',
+	options: { page?: number; limit?: number; search?: string } = {}
+) => {
+	let error = null;
+
+	const { page = 1, limit = 10, search } = options;
+	const params = new URLSearchParams({
+		page: page.toString(),
+		limit: limit.toString()
+	});
+
+	if (search && search.trim()) {
+		params.append('search', search.trim());
+	}
+
+	const res = await fetch(`${WEBUI_API_BASE_URL}/evaluations/feedbacks/all/paginated?${params}`, {
+		method: 'GET',
+		headers: {
+			Accept: 'application/json',
+			'Content-Type': 'application/json',
+			authorization: `Bearer ${token}`
+		}
+	})
+		.then(async (res) => {
+			if (!res.ok) throw await res.json();
+			return res.json();
+		})
+		.then((json) => {
+			return json;
+		})
+		.catch((err) => {
+			error = err.detail;
+			console.log(err);
+			return null;
+		});
+
+	if (error) {
+		throw error;
+	}
+
+	return res;
+};
+
+export const getFeedbacksCount = async (token: string = '', search?: string) => {
+	let error = null;
+
+	const params = new URLSearchParams();
+	if (search && search.trim()) {
+		params.append('search', search.trim());
+	}
+
+	const res = await fetch(`${WEBUI_API_BASE_URL}/evaluations/feedbacks/count?${params}`, {
+		method: 'GET',
+		headers: {
+			Accept: 'application/json',
+			'Content-Type': 'application/json',
+			authorization: `Bearer ${token}`
+		}
+	})
+		.then(async (res) => {
+			if (!res.ok) throw await res.json();
+			return res.json();
+		})
+		.then((json) => {
+			return json;
+		})
+		.catch((err) => {
+			error = err.detail;
+			console.log(err);
+			return null;
+		});
+
+	if (error) {
+		throw error;
+	}
+
+	return res;
+};
+
+// Legacy function for backward compatibility (now calling original endpoint)
+export const getAllFeedbacksLegacy = async (token: string = '') => {
 	let error = null;
 
 	const res = await fetch(`${WEBUI_API_BASE_URL}/evaluations/feedbacks/all`, {