@pascal-beaulne 2025-07-11
I've double checked with a normal user and it's taking forever for them to load prompts like it is for us admins. While we're here can we either implement some type of lazy loading or make better queries to mitigate it?